+++ hutool-core/src/test/java/com/xiaoleilu/hutool/util/StrUtilTest.java
@@ -101,4 +101,34 @@ public class StrUtilTest {
 		strip = StrUtil.stripIgnoreCase(str, null, "567");
 		Assert.assertEquals("abcd123", strip);
 	}
+	
+	@Test
+	public void indexOfIgnoreCaseTest() {
+		Assert.assertEquals(-1, StrUtil.indexOfIgnoreCase(null, "balabala", 0));
+		Assert.assertEquals(-1, StrUtil.indexOfIgnoreCase("balabala" , null, 0));
+		Assert.assertEquals(0, StrUtil.indexOfIgnoreCase("", "", 0));
+		Assert.assertEquals(0, StrUtil.indexOfIgnoreCase("aabaabaa", "A", 0));
+		Assert.assertEquals(2 , StrUtil.indexOfIgnoreCase("aabaabaa", "B", 0));
+		Assert.assertEquals(1 , StrUtil.indexOfIgnoreCase("aabaabaa", "AB", 0));
+		Assert.assertEquals(5 , StrUtil.indexOfIgnoreCase("aabaabaa", "B", 3));
+		Assert.assertEquals(-1 , StrUtil.indexOfIgnoreCase("aabaabaa", "B", 9));
+		Assert.assertEquals(2 , StrUtil.indexOfIgnoreCase("aabaabaa", "B", -1));
+		Assert.assertEquals(2 , StrUtil.indexOfIgnoreCase("aabaabaa", "", 2));
+		Assert.assertEquals(-1 , StrUtil.indexOfIgnoreCase("abc", "", 9));
+	}
+	
+	@Test
+	public void lastIndexOfIgnoreCaseTest() {
+		Assert.assertEquals(-1, StrUtil.lastIndexOfIgnoreCase(null, "balabala", 0));
+		Assert.assertEquals(-1, StrUtil.lastIndexOfIgnoreCase("balabala" , null));
+		Assert.assertEquals(0, StrUtil.lastIndexOfIgnoreCase("", ""));
+		Assert.assertEquals(7, StrUtil.lastIndexOfIgnoreCase("aabaabaa", "A"));
+		Assert.assertEquals(5 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "B"));
+		Assert.assertEquals(4 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "AB"));
+		Assert.assertEquals(2 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "B", 3));
+		Assert.assertEquals(5 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "B", 9));
+		Assert.assertEquals(-1 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "B", -1));
+		Assert.assertEquals(2 , StrUtil.lastIndexOfIgnoreCase("aabaabaa", "", 2));
+		Assert.assertEquals(3 , StrUtil.lastIndexOfIgnoreCase("abc", "", 9));
+	}
 }